% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ALS.R
\name{read_ALS}
\alias{read_ALS}
\title{Automatic reading of ALS reports in PDF-format}
\usage{
read_ALS(file)
}
\arguments{
\item{file}{the name of the ALS report to be read from.}
}
\description{
\code{read_ALS} reads a an ALS report in PDF format and returns a S6 object
containing all the tables
}
