% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ALS.R
\name{reports}
\alias{reports}
\title{Returns all reports found in the ALS file in one data frame}
\usage{
reports(alsreport)
}
\arguments{
\item{alsreport}{the name of the ALS object to return the reports from}
}
\description{
\code{reports(als)} Returns the reports in the parsed als object
}
